Update MQTT Connector and tests
