+++ cloud_connectors/azure/mqtt_connector/res/azure_function/src/function.csproj
