+++ cloud_connectors/azure/mqtt_connector/res/azure_function/tests/MQTTConnectorAzureFunction.Tests.csproj
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\src\MQTTConnector.csproj" />
+    <ProjectReference Include="..\src\function.csproj" />
   </ItemGroup>
 
 </Project>
